+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/TeaApplication.java
@@ -131,6 +131,10 @@ public class TeaApplication implements Application, Runnable {
         initGdx();
         initSound();
 
+        if(config.preloadListener != null) {
+            config.preloadListener.onPreload(assetLoader);
+        }
+
         Gdx.app = this;
         Gdx.graphics = graphics;
         Gdx.gl = graphics.getGL20();